functions/zoom_oauth_callback/requirements.txt
functions-framework
requests
google-cloud-secret-manager
